Sprint4 partial merge

+++ server/server.js
@@ -7,6 +7,7 @@ const passport = require("passport");
 const LocalStrategy = require("passport-local").Strategy;
 const session = require("express-session");
 const dao = require("./dao"); // module for accessing the DB
+const telegramBot = require("./telegrambot/SendMessage.js");
 //Per validazione aggiuntiva
 const validator = require("validator");
 let testmode = false;
@@ -67,6 +68,24 @@ passport.use(
   new LocalStrategy(function (username, password, done) {
     dao
       .getWarehouseWorker(username, password)
+      .then((user) => {
+        if (!user)
+          return done(null, false, {
+            message: "Incorrect username and/or password",
+          });
+        return done(null, user);
+      })
+      .catch((err) => {
+        return done(err);
+      });
+  })
+);
+
+passport.use(
+  "warehouse-manager-local",
+  new LocalStrategy(function (username, password, done) {
+    dao
+      .getWarehouseManager(username, password)
       .then((user) => {
         if (!user)
           return done(null, false, {
@@ -165,6 +184,16 @@ passport.deserializeUser((id, done) => {
   }
   if (type === "M") {
     dao
+      .getWarehouseManagerById(identifier)
+      .then((user) => {
+        done(null, user);
+      })
+      .catch((err) => {
+        done(err, null);
+      });
+  }
+  if (type === "A") {
+    dao
       .getManagerById(identifier)
       .then((user) => {
         done(null, user);
@@ -249,7 +278,26 @@ app.post("/api/warehouseWorkerSessions", function (req, res, next) {
   })(req, res, next);
 });
 
-//POST /api/managerSessions FOR LOGIN OF CLIENT
+//POST /api/warehouseManagerSessions FOR LOGIN OF WAREHOUSE MANGER
+app.post("/api/warehouseManagerSessions", function (req, res, next) {
+  passport.authenticate("warehouse-manager-local", (err, user, info) => {
+    if (err) return next(err);
+    if (!user) {
+      // display wrong login messages
+      return res.status(401).json(info);
+    }
+    // success, perform the login
+    req.login(user, (err) => {
+      if (err) return next(err);
+
+      // req.user contains the authenticated user, we send all the user info back
+      // this is coming from userDao.getUser()
+      return res.json(req.user);
+    });
+  })(req, res, next);
+});
+
+//POST /api/warehouseManagerSessions FOR LOGIN OF MANGER
 app.post("/api/managerSessions", function (req, res, next) {
   passport.authenticate("manager-local", (err, user, info) => {
     if (err) return next(err);
@@ -1004,14 +1052,15 @@ app.put("/api/walletbalance", isLoggedIn, async (req, res) => {
     amount: req.body.amount,
   };
 
-  dao
-    .updateWallet(wallet)
-    .then(() => {
-      res.status(200).json(wallet);
-    })
-    .catch((error) => {
+  try{
+    await dao.updateWallet(wallet);
+    await telegramBot.SendMessage(wallet.id,`Your wallet was modified. New balance : ${wallet.amount} €`);
+    res.status(200).json(wallet);
+  }
+  catch(error){
       res.status(500).json(error);
-    });
+  }
+
 });
 
 // GET /api/bookingsPendingCancelation to get all bookings with PENDINGCANCELATION state
@@ -1258,6 +1307,11 @@ let clockDate = new Date();
 let timers = setInterval(async () => {clockDate = new Date();
   await clockActions();
 }, 30000);
+//FLAG to execute queries once a day, resetted every sunday
+let once = [true,true,true,true,true,true,true];
+
+//For periodic remainder, every 4 hours
+let oldHours =0;
 
 //GET /api/time to get current time
 app.get("/api/time",  (req, res) => {
@@ -1300,45 +1354,66 @@ async function clockActions(){
     /* On TUESDAY farmers have delivered their products. Now it's time to check if all the bookings that are 
     BOOKED should become CONFIRMED or PENDINGCANCELATION. We should keep in the booking only the products
     CONFIRMED BY FARMERS. If all products of a booking are not confirmed the booking become EMPTY 
-    and the client is notified. */
+    and the client is notified. 
+    Periodic telegram message is sent if customer needs to top up
+    A message is sent if booking is confirmed*/
     
     if(clockDate.getDay()===2){
 
-      //Delete from bookings all product still expected, so unconfirmed
-      await dao.deleteBookingProductsExpected();
+      if(once[2]){
+        //Delete from bookings all product still expected, so unconfirmed
+        await dao.deleteBookingProductsExpected();
 
-      const bookings = await dao.getTotal();
-      
-      bookings.forEach(async (booking)=>{
-        //Get the wallet  of the customer and put in variable wallet
-        const wallet = await dao.getWallet(booking.client);
-        if (wallet.balance >= booking.total) {
-          //PUT in state CONFIRMED
-          await dao.editStateBooking({id: booking.id, state: "CONFIRMED"});
-          //Update amount
-          await dao.updateWallet({ amount: wallet.balance - booking.total, id: booking.client });
+        const bookings = await dao.getTotal();
+        
+        bookings.forEach(async (booking)=>{
+          //Get the wallet  of the customer and put in variable wallet
+          const wallet = await dao.getWallet(booking.client);
+          if (wallet.balance >= booking.total) {
+            //PUT in state CONFIRMED
+            await dao.editStateBooking({id: booking.id, state: "CONFIRMED"});
+            //Update amount
+            await dao.updateWallet({ amount: wallet.balance - booking.total, id: booking.client });
+            //Send telegram message
+            const bookingProducts = await dao.productsOfBooking(booking.id);
+            telegramBot.SendMessage(booking.client,`Purchase confirmation, booking #${booking.id}:\n\n${bookingProducts.map((p)=>p.qty +" " +p.product)}\n\nTotal: ${booking.total} €`);
+          } 
+          else {
+            //Put in state PENDINGCANCELATION
+            await dao.editStateBooking({ id: booking.id, state: "PENDINGCANCELATION" });
+          }
 
-        } 
-        else {
-          //Put in state PENDINGCANCELATION
-          await dao.editStateBooking({ id: booking.id, state: "PENDINGCANCELATION" });
-        }
 
+        });
 
-      });
-
-      //If empty bookings put in state EMPTY
-      const emptyBknings= await dao.getEmptyBookings();
-      
-      emptyBknings.forEach(async (booking)=>{
-        await dao.editStateBooking({ id: booking.id, state: "EMPTY" });
-      });
+        //If empty bookings put in state EMPTY
+        const emptyBknings= await dao.getEmptyBookings();
+        
+        emptyBknings.forEach(async (booking)=>{
+          await dao.editStateBooking({ id: booking.id, state: "EMPTY" });
+        });
+        once[2]=false;
+      }
+      //Every 4 hour remember to top up to all clients that have bookings in pendingcancelation state
+      if(clockDate.getHours()-oldHours>=4){
+        oldHours=clockDate.getHours()-oldHours;
+        const bookings = await dao.getTotalPendingCancelation();
+        let alreadySent =[];
+        bookings.forEach(async (booking)=>{
+          if(!alreadySent.includes(booking.client)){
+            const wallet = await dao.getWallet(booking.client);
+            telegramBot.SendMessage(booking.client,`Your current balance (${wallet.balance} €) is insufficient to complete the order #${booking.id}. Please top-up at least ${booking.total-wallet.balance}€ to complete the order.`);
+            alreadySent.push(booking.client);
+          }
+        })
+      }
     }
 
     /* Until WEDNESDAY customer have the possibility to top up their wallets if they have orders
     in state PENDINGCANCELATION. If they do and the balance is enough, their orders will return in
-    CONFIRMED state, otherwise they will be CANCELED. */
-    if(clockDate.getDay()===3){
+    CONFIRMED state, otherwise they will be CANCELED. 
+    A message is sent if booking is confirmed*/
+    if(clockDate.getDay()===3 && once[3]){
 
       const bookings = await dao.getTotalPendingCancelation();
       bookings.forEach(async (booking)=>{
@@ -1347,6 +1422,9 @@ async function clockActions(){
         if (wallet.balance >= booking.total) {
           //PUT in state CONFIRMED
           await dao.editStateBooking({id: booking.id, state: "CONFIRMED"});
+          //Send telegram message
+          const bookingProducts = await dao.productsOfBooking(booking.id);
+          telegramBot.SendMessage(booking.client,`Purchase confirmation, booking #${booking.id}:\n\n${bookingProducts.map((p)=>p.qty +" " +p.product)}\n\nTotal: ${booking.total} €`);
           //Update amount
           await dao.updateWallet({ amount: wallet.balance - booking.total, id: booking.client });
 
@@ -1359,15 +1437,17 @@ async function clockActions(){
       });
 
 
-
+      once[3]=false;
     }
 
     /* On SATURDAY MORNING a new week starts. So we move bookings from booking table to booking history
     table. If a booking was in COMPLETED or EMPTY or CANCELED state, it will be saved in booking history as COMPLETED
     or EMPTY or CANCELED. If it was in CONFIRMED state and Delivery Mode was pickup, it will be saved as
     UNRETRIEVED. If it was CONFIRMED and Delivery mode was delivery, we just delete it from booking, we don't care
-    about this state now (it's for future stories). */
-    if(clockDate.getDay()===6){
+    about this state now (it's for future stories). 
+    
+    Moreover, we move all products to EXPECTED state and we add 10 to every quantity*/
+    if(clockDate.getDay()===6 && once[6]){
       
       const startDate = new Date(clockDate);
       //If this happens only on saturday, starday will be monday
@@ -1412,9 +1492,26 @@ async function clockActions(){
         }
 
       })
+
+      //Move all products in product_week to expected and increase qty by 10
+      await dao.resetProductWeekVC();
+
+      once[6]=false;
+      once[1]=true;
     }
 
-
+    //On sunday set all once to true
+    if(clockDate.getDay()===0 && once[0]){
+      for (let i=1;i<7;i++){
+        once[i]=true;
+      }
+      const clients= await dao.getClients();
+      //Send a message to every customer on sunday morning
+      clients.forEach((c)=>{
+        telegramBot.SendMessage(c.id.substring(1),"Updated list of available products is now available on website.");
+      });
+      once[0]=false;
+    }
 
   }
